pyproject.toml
[tool.poetry]
name = "yamless"
version = "0.1.0"
description = ""
authors = ["Thunder <mahmmoud.hassanein@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
docker = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
